strings to numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,7 +8,7 @@
  */
 int _atoi(char *s)
 {
-int res = 0;
+unsigned long int res = 0;
 int sign = 1;
 int i = 0;
 if (s[0] == '-')
